Added functionality Integer and Range kind of Class Nodes

+++ de.ovgu.featureide.fm.ui/src/de/ovgu/featureide/fm/ui/editors/FeatureDiagramEditor.java
@@ -373,7 +373,8 @@ public class FeatureDiagramEditor extends ScrollingGraphicalViewer implements
 			menu.add(createConstraintAction);
 			menu.add(renameAction);
 			//menu.add(deleteAction); //Abhi
-			menu.add(deleteAllAction);
+			if(selectionAction.isNodeRangeClass == false)
+				menu.add(deleteAllAction);
 			menu.add(new Separator());
 			//connectionEntrys(menu);//Abhi
 			//menu.add(mandatoryAction); //Abhi
